requirements.txt
starlette>=0.13.2
aiomysql>=0.0.0
databases>=0.2.6
discord.py>1.3.2
uvicorn>=0.11.3
sqlalchemy>=1.3.13
aiohttp>=3.6.2
discord.py>=1.0.0
webargs-starlette>=1.2.0
git+https://github.com/WardPearce/aiob2.git
git+https://github.com/MertYakupBaykan/aioproxyio.git
git+https://github.com/WardPearce/aiodactyl.git
python-multipart